plugins/lockstep: made the socket path argument not positional
Socket name was a positional argument that must be provided as the first
argument. It now can be provided as a value to the argument "sockpath".
It follows the same convention of taking the last-provided value of an
argument and ignore other values.

Also, on failure, printed the error report to stderr instead of
qemu_plugin_outs.

Signed-off-by: Mahmoud Mandour <ma.mandourr@gmail.com>